fix: プラグインの note:new/notification:new fan-out を復旧し dead code を削除
registerQuery の唯一の呼び出し元が dead code の useQuerySubscription
だったため、queryId -> flavor マップが常に空で AiScript プラグインの
Nd:on('note:new'/'notification:new') が全 delta を捨てていた。

- レジストリを core/queryRegistry に分離 (adapters → aiscript → stores
  → adapters の循環 import 回避) し、実稼働経路の
  createQuerySubscription が登録/解除する
- events.ts の独自 queryDelta.listen を queryDeltaBus 経由に統合
  (復帰時再アタッチの恩恵を受ける)
- dead code を削除: useQuerySubscription.ts (呼び出し元ゼロ)、
  destroyAllAdapters (同)
