+++ src/adapters/factory.ts
@@ -93,18 +93,6 @@ export function destroyAdapter(accountId: string): void {
   adapterCache.delete(accountId)
 }
 
-/**
- * Destroy all cached adapters.
- * Call on app shutdown or full reset.
- */
-export function destroyAllAdapters(): void {
-  for (const adapter of adapterCache.values()) {
-    adapter.stream.cleanup()
-  }
-  adapterCache.clear()
-  adapterPending.clear()
-}
-
 /** Check if an adapter is cached for the given accountId (for testing/debug) */
 export function hasAdapterCache(accountId: string): boolean {
   return adapterCache.has(accountId)
